requirements.txt
# Mobile Comment Generator - 基本依存関係
# 互換性を重視したバージョン指定

# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0

# LangGraph ecosystem - 互換性のあるバージョン組み合わせ
langgraph>=0.0.35
langchain>=0.1.0
langchain-core>=0.1.42  # langgraph 0.0.35 の要求に合わせて更新

# LLM Providers - 柔軟なバージョン指定
openai>=1.12.0
google-generativeai>=0.3.2
anthropic>=0.18.1

# AWS Integration (No longer needed - using local CSV files)
# boto3>=1.34.0

# HTTP/API Communication
requests>=2.31.0
aiohttp>=3.9.1

# Data Processing
pandas>=2.1.4
numpy>=1.26.2
scikit-learn>=1.3.2

# UI Framework
streamlit>=1.28.0  # AppTest framework対応

# Utilities
pyperclip>=1.8.2    # クリップボード機能
pytz>=2023.3        # タイムゾーン処理
loguru>=0.7.2       # ログ管理